+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Changelog
 
-- Pequena edição em `./util/embed.js` para facilitação e otimização no código
-- Edição de todos arquivos que utilizavam a versão antiga do *PixelEmbed* para a nova versão
+- Separado eventos em sub-categorias do `./events/InteractionCreate.js`
+- `./events/Button.js` criado
+- `./events/ModalSubmit.js` criado
+- `./events/StringSelectMenu.js` criado